Unicode handling fixed; test_socketlevel split up
