+++ test/with_dummyserver/test_socketlevel.py
@@ -1,6 +1,3 @@
-# TODO: Break this module up into pieces. Maybe group by functionality tested
-# rather than the socket level-ness of it.
-
 from urllib3 import HTTPConnectionPool, HTTPSConnectionPool
 from urllib3.poolmanager import proxy_from_url
 from urllib3.exceptions import (
@@ -11,7 +8,6 @@ from urllib3.exceptions import (
         SSLError,
         ProtocolError,
 )
-from urllib3.response import httplib
 from urllib3.util.ssl_ import HAS_SNI
 from urllib3.util.timeout import Timeout
 from urllib3.util.retry import Retry
@@ -21,41 +17,12 @@ from dummyserver.testcase import SocketDummyServerTestCase
 from dummyserver.server import (
     DEFAULT_CERTS, DEFAULT_CA, get_unreachable_address)
 
-from .. import onlyPy3, LogRecorder
-
 from nose.plugins.skip import SkipTest
-try:
-    from mimetools import Message as MimeToolMessage
-except ImportError:
-    class MimeToolMessage(object):
-        pass
+
 from threading import Event
 import select
 import socket
 import ssl
-
-
-class TestCookies(SocketDummyServerTestCase):
-
-    def test_multi_setcookie(self):
-        def multicookie_response_handler(listener):
-            sock = listener.accept()[0]
-
-            buf = b''
-            while not buf.endswith(b'\r\n\r\n'):
-                buf += sock.recv(65536)
-
-            sock.send(b'HTTP/1.1 200 OK\r\n'
-                      b'Set-Cookie: foo=1\r\n'
-                      b'Set-Cookie: bar=1\r\n'
-                      b'\r\n')
-            sock.close()
-
-        self._start_server(multicookie_response_handler)
-        pool = HTTPConnectionPool(self.host, self.port)
-        r = pool.request('GET', '/', retries=0)
-        self.assertEqual(r.headers, {'set-cookie': 'foo=1, bar=1'})
-        self.assertEqual(r.headers.getlist('set-cookie'), ['foo=1', 'bar=1'])
 
 
 class TestSNI(SocketDummyServerTestCase):
@@ -721,100 +688,6 @@ class TestErrorWrapping(SocketDummyServerTestCase):
         pool = HTTPConnectionPool(self.host, self.port, retries=False)
         self.assertRaises(ProtocolError, pool.request, 'GET', '/')
 
-class TestHeaders(SocketDummyServerTestCase):
-
-    @onlyPy3
-    def test_httplib_headers_case_insensitive(self):
-        self.start_response_handler(
-           b'HTTP/1.1 200 OK\r\n'
-           b'Content-Length: 0\r\n'
-           b'Content-type: text/plain\r\n'
-           b'\r\n'
-        )
-        pool = HTTPConnectionPool(self.host, self.port, retries=False)
-        HEADERS = {'Content-Length': '0', 'Content-type': 'text/plain'}
-        r = pool.request('GET', '/')
-        self.assertEqual(HEADERS, dict(r.headers.items()))  # to preserve case sensitivity
-
-    def test_headers_are_sent_with_the_original_case(self):
-        headers = {'foo': 'bar', 'bAz': 'quux'}
-        parsed_headers = {}
-
-        def socket_handler(listener):
-            sock = listener.accept()[0]
-
-            buf = b''
-            while not buf.endswith(b'\r\n\r\n'):
-                buf += sock.recv(65536)
-
-            headers_list = [header for header in buf.split(b'\r\n')[1:] if header]
-
-            for header in headers_list:
-                (key, value) = header.split(b': ')
-                parsed_headers[key.decode()] = value.decode()
-
-            # Send incomplete message (note Content-Length)
-            sock.send((
-                'HTTP/1.1 204 No Content\r\n'
-                'Content-Length: 0\r\n'
-                '\r\n').encode('utf-8'))
-
-            sock.close()
-
-        self._start_server(socket_handler)
-        expected_headers = {'Accept-Encoding': 'identity',
-                            'Host': '{0}:{1}'.format(self.host, self.port)}
-        expected_headers.update(headers)
-
-        pool = HTTPConnectionPool(self.host, self.port, retries=False)
-        pool.request('GET', '/', headers=HTTPHeaderDict(headers))
-        self.assertEqual(expected_headers, parsed_headers)
-
-
-class TestBrokenHeaders(SocketDummyServerTestCase):
-    def setUp(self):
-        if issubclass(httplib.HTTPMessage, MimeToolMessage):
-            raise SkipTest('Header parsing errors not available')
-
-        super(TestBrokenHeaders, self).setUp()
-
-    def _test_broken_header_parsing(self, headers):
-        self.start_response_handler((
-           b'HTTP/1.1 200 OK\r\n'
-           b'Content-Length: 0\r\n'
-           b'Content-type: text/plain\r\n'
-           ) + b'\r\n'.join(headers) + b'\r\n'
-        )
-
-        pool = HTTPConnectionPool(self.host, self.port, retries=False)
-
-        with LogRecorder() as logs:
-            pool.request('GET', '/')
-
-        for record in logs:
-            if 'Failed to parse headers' in record.msg and \
-                    pool._absolute_url('/') == record.args[0]:
-                return
-        self.fail('Missing log about unparsed headers')
-
-    def test_header_without_name(self):
-        self._test_broken_header_parsing([
-            b': Value\r\n',
-            b'Another: Header\r\n',
-        ])
-
-    def test_header_without_name_or_value(self):
-        self._test_broken_header_parsing([
-            b':\r\n',
-            b'Another: Header\r\n',
-        ])
-
-    def test_header_without_colon_or_value(self):
-        self._test_broken_header_parsing([
-            b'Broken Header',
-            b'Another: Header',
-        ])
-
 
 class TestHEAD(SocketDummyServerTestCase):
     def test_chunked_head_response_does_not_hang(self):
@@ -843,66 +716,3 @@ class TestHEAD(SocketDummyServerTestCase):
         # stream will use the read method here.
         self.assertEqual([], list(r.stream()))
 
-
-class TestChunkedTransfer(SocketDummyServerTestCase):
-    def _start_chunked_handler(self):
-        self.buffer = b''
-
-        def socket_handler(listener):
-            sock = listener.accept()[0]
-
-            while not self.buffer.endswith(b'\r\n0\r\n\r\n'):
-                self.buffer += sock.recv(65536)
-
-            sock.send(
-               b'HTTP/1.1 200 OK\r\n'
-               b'Content-type: text/plain\r\n'
-               b'Content-Length: 0\r\n'
-               b'\r\n')
-            sock.close()
-
-        self._start_server(socket_handler)
-
-    def test_chunks(self):
-        self._start_chunked_handler()
-        chunks = ['foo', 'bar', '', 'bazzzzzz']
-        pool = HTTPConnectionPool(self.host, self.port, retries=False)
-        r = pool.urlopen('GET', '/', chunks, headers=dict(DNT='1'), chunked=True)
-
-        self.assertTrue(b'Transfer-Encoding' in self.buffer)
-        body = self.buffer.split(b'\r\n\r\n', 1)[1]
-        lines = body.split(b'\r\n')
-        # Empty chunks should have been skipped, as this could not be distinguished
-        # from terminating the transmission
-        for i, chunk in enumerate([c for c in chunks if c]):
-            self.assertEqual(lines[i * 2], str(len(chunk)).encode('utf-8'))
-            self.assertEqual(lines[i * 2 + 1], chunk.encode('utf-8'))
-
-    def test_string_body(self):
-        self._start_chunked_handler()
-        pool = HTTPConnectionPool(self.host, self.port, retries=False)
-        body_str = 'thisshouldbeonechunk'
-        r = pool.urlopen('GET', '/', body_str, chunked=True)
-        header, body = self.buffer.split(b'\r\n\r\n', 1)
-
-        self.assertTrue(b'Transfer-Encoding: chunked' in header.split(b'\r\n'))
-        self.assertTrue(b'\r\n' + body_str.encode('utf-8') + b'\r\n' in body)
-        self.assertTrue(body.endswith(b'\r\n0\r\n\r\n'))
-
-    def test_no_body(self):
-        self._start_chunked_handler()
-        pool = HTTPConnectionPool(self.host, self.port, retries=False)
-        r = pool.urlopen('GET', '/', None, chunked=True)
-        header, body = self.buffer.split(b'\r\n\r\n', 1)
-
-        self.assertTrue(b'Transfer-Encoding: chunked' in header.split(b'\r\n'))
-        self.assertEqual(body, b'0\r\n\r\n')
-
-    def test_empty_iterable_body(self):
-        self._start_chunked_handler()
-        pool = HTTPConnectionPool(self.host, self.port, retries=False)
-        r = pool.urlopen('GET', '/', [], chunked=True)
-        header, body = self.buffer.split(b'\r\n\r\n', 1)
-
-        self.assertTrue(b'Transfer-Encoding: chunked' in header.split(b'\r\n'))
-        self.assertEqual(body, b'0\r\n\r\n')
